# openapi/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "API local de pedidos",
    "version": "1.0.0",
    "description": "Ejercicio individual. Importes en quetzales. Preparación exclusivamente local; no usar en producción. El listado filtra por propietario. POST no es idempotente."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Solo loopback"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/orders": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "operationId": "listOrders",
        "summary": "Listar pedidos propios con paginación",
        "description": "Orden de inserción. Se rechazan parámetros desconocidos y duplicados. Una página fuera del total devuelve items vacío. No acepta cuerpo.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Página desde 1",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000000,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Tamaño de página",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20,
              "default": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página propia",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPage"
                }
              }
            }
          },
          "400": {
            "description": "Datos, UUID o parámetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciales ausentes o inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        }
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "Bearer"
                  ]
                },
                "description": "Desafío Bearer"
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pedidos"
        ],
        "operationId": "createOrder",
        "summary": "Crear borrador",
        "description": "Precio fijo Q25 por CUADERNO; total=quantity*25. No acepta owner ni total enviados por cliente. Cada POST válido crea un pedido nuevo.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrder"
              },
              "example": {
                "product": "CUADERNO",
                "quantity": 2
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            },
            "headers": {
              "Location": {
                "required": true,
                "description": "URI relativa del pedido creado",
                "schema": {
                  "type": "string",
                  "pattern": "^/orders/[0-9a-f-]{36}$"
                }
              }
            }
          },
          "400": {
            "description": "Datos, UUID o parámetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciales ausentes o inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        }
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "Bearer"
                  ]
                },
                "description": "Desafío Bearer"
              }
            }
          },
          "413": {
            "description": "Cuerpo mayor de 16 KiB",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "PAYLOAD_TOO_LARGE"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "Cuerpo debe usar application/json",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNSUPPORTED_MEDIA_TYPE"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/__demo/prepare": {
      "post": {
        "tags": [
          "Preparación"
        ],
        "security": [],
        "operationId": "prepareDemo",
        "summary": "Reiniciar fixtures locales y emitir sesiones temporales",
        "description": "Destructivo únicamente para los datos sintéticos de esta demo. Solo servidor loopback. Rechaza encabezado Origin. No acepta datos personales. No publicar el servicio ni exportar el cuerpo de esta respuesta.",
        "responses": {
          "200": {
            "description": "Fixtures de esta ejecución",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fixture"
                }
              }
            }
          },
          "403": {
            "description": "El pedido pertenece a otro usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "FORBIDDEN"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Opaque demo session",
        "description": "Sesión aleatoria local obtenida durante preparación, sin contraseñas"
      }
    },
    "schemas": {
      "CreateOrder": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "product",
          "quantity"
        ],
        "properties": {
          "product": {
            "type": "string",
            "enum": [
              "CUADERNO"
            ]
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          }
        }
      },
      "Order": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "owner",
          "product",
          "quantity",
          "total",
          "status",
          "confirmationCount",
          "createdAt",
          "confirmedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "owner": {
            "type": "string",
            "enum": [
              "usuario_a",
              "usuario_b"
            ]
          },
          "product": {
            "type": "string",
            "enum": [
              "CUADERNO"
            ]
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "total": {
            "type": "number",
            "minimum": 25,
            "maximum": 250,
            "multipleOf": 25
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "confirmed"
            ]
          },
          "confirmationCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "confirmedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "OrderPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "page",
          "limit",
          "total",
          "totalPages"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "Fixture": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "users",
          "tokens",
          "ownIds",
          "foreignId"
        ],
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "usuario_a",
                "usuario_b"
              ]
            },
            "minItems": 2,
            "maxItems": 2
          },
          "tokens": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "usuario_a",
              "usuario_b"
            ],
            "properties": {
              "usuario_a": {
                "type": "string",
                "pattern": "^[a-f0-9]{64}$"
              },
              "usuario_b": {
                "type": "string",
                "pattern": "^[a-f0-9]{64}$"
              }
            }
          },
          "ownIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            },
            "minItems": 3,
            "maxItems": 3
          },
          "foreignId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          }
        }
      }
    }
  }
}